Fix Trilean.ToSignedInteger to return sbyte, preserving Na sentinel
The method returned byte, so the Na sentinel (-1 from ToSignedNumber<int>()) wrapped to 255 under the byte cast — callers could not distinguish Na from "a very large number" despite the method name promising a signed integer.

+++ src/DSE.Open.Abstractions/Trilean.cs
@@ -165,9 +165,15 @@ public readonly struct Trilean
         };
     }
 
-    public byte ToSignedInteger()
+    /// <summary>
+    /// Converts the current value to a signed 8-bit integer.
+    /// </summary>
+    /// <returns>
+    /// <c>1</c> if <see cref="IsTrue"/>, <c>0</c> if <see cref="IsFalse"/>, or <c>-1</c> if <see cref="IsNa"/>.
+    /// </returns>
+    public sbyte ToSignedInteger()
     {
-        return (byte)ToSignedNumber<int>();
+        return ToSignedNumber<sbyte>();
     }
 
     public bool ToBoolean()
